functions/requirements.txt
firebase_functions~=0.1.0
tensorflow>=2.12.0
Pillow>=10.0.0
numpy>=1.21.0